+++ plate.php
@@ -23,8 +23,8 @@ declare(strict_types=1);
 add_action('after_setup_theme', function () {
     require_if_theme_supports('plate-dashboard', __DIR__.'/src/dashboard.php');
     require_if_theme_supports('plate-editor', __DIR__.'/src/editor.php');
-    require_if_theme_supports('plate-footer', __DIR__.'/src/footer.php');
-    require_if_theme_supports('plate-login', __DIR__.'/src/login.php');
+    require_if_theme_supports('plate-footer-text', __DIR__.'/src/footer-text.php');
+    require_if_theme_supports('plate-login-logo', __DIR__.'/src/login-logo.php');
     require_if_theme_supports('plate-menu', __DIR__.'/src/menu.php');
     require_if_theme_supports('plate-permalink', __DIR__.'/src/permalink.php');
     require_if_theme_supports('plate-tabs', __DIR__.'/src/tabs.php');