+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 4.0.0 (upcoming)
 
+- Rename plate-footer to plate-footer-text
+- Rename plate-login to plate-login-logo
+
 ## 3.1.0 (released 2018-02-20)
 
 - Added support for sub-menus